+++ .github/workflows/argon2.yml
@@ -37,7 +37,7 @@ jobs:
           targets: ${{ matrix.target }}
       - run: cargo build --target ${{ matrix.target }} --no-default-features
       - run: cargo build --target ${{ matrix.target }} --no-default-features --features password-hash
-      - run: cargo build --target ${{ matrix.target }} --no-default-features --features simple
+      - run: cargo build --target ${{ matrix.target }} --no-default-features --features password-hash
       - run: cargo build --target ${{ matrix.target }} --no-default-features --features zeroize
 
   minimal-versions:
