+++ .readme/Cargo.toml
@@ -8,5 +8,5 @@ publish = false
 [dependencies]
 password-hash = "0.6.0-rc.3"
 argon2 = { path = "../argon2" }
-pbkdf2 = { path = "../pbkdf2", features = ["simple"] }
+pbkdf2 = { path = "../pbkdf2", features = ["password-hash"] }
 scrypt = { path = "../scrypt" }
