+++ Cargo.toml
@@ -20,3 +20,5 @@ opt-level = 2
 argon2 = { path = "./argon2" }
 pbkdf2 = { path = "./pbkdf2" }
 scrypt = { path = "./scrypt" }
+
+password-hash = { git = "https://github.com/RustCrypto/traits" }
