+++ argon2/Cargo.toml
@@ -23,7 +23,6 @@ blake2 = { version = "0.11.0-rc.3", default-features = false }
 # optional dependencies
 rayon = { version = "1.7", optional = true }
 password-hash = { version = "0.6.0-rc.5", optional = true, features = ["phc"] }
-phc = { version = "0.6.0-rc.0", optional = true, features = ["rand_core"] }
 zeroize = { version = "1", default-features = false, optional = true }
 
 [target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
@@ -33,12 +32,13 @@ cpufeatures = "0.2.17"
 hex-literal = "1"
 
 [features]
-default = ["alloc", "getrandom", "simple"]
+default = ["alloc", "getrandom", "password-hash"]
 alloc = ["password-hash?/alloc"]
 
-getrandom = ["simple", "password-hash/getrandom"]
+getrandom = ["password-hash/getrandom"]
 parallel = ["dep:rayon"]
-simple = ["password-hash", "phc"]
+password-hash = ["dep:password-hash"]
+rand_core = ["password-hash/rand_core"]
 zeroize = ["dep:zeroize"]
 
 [lints.rust.unexpected_cfgs]
