+++ pbkdf2/Cargo.toml
@@ -19,7 +19,6 @@ digest = { version = "0.11.0-rc.4", features = ["mac"] }
 # optional dependencies
 hmac = { version = "0.13.0-rc.3", default-features = false, optional = true }
 password-hash = { version = "0.6.0-rc.5", default-features = false, optional = true, features = ["phc"] }
-phc = { version = "0.6.0-rc.0", optional = true, features = ["rand_core"] }
 sha1 = { version = "0.11.0-rc.3", default-features = false, optional = true }
 sha2 = { version = "0.11.0-rc.3", default-features = false, optional = true }
 
@@ -33,8 +32,9 @@ belt-hash = "0.2.0-rc.3"
 
 [features]
 default = ["hmac"]
-getrandom = ["simple", "password-hash/getrandom"]
-simple = ["hmac", "dep:password-hash", "dep:phc", "sha2"]
+getrandom = ["password-hash", "password-hash/getrandom"]
+password-hash = ["hmac", "dep:password-hash", "sha2"]
+rand_core = ["password-hash/rand_core"]
 
 [package.metadata.docs.rs]
 all-features = true
