+++ scrypt/src/phc.rs
