+++ yescrypt/src/error.rs
@@ -37,7 +37,7 @@ impl From<TryFromIntError> for Error {
     }
 }
 
-#[cfg(feature = "simple")]
+#[cfg(feature = "password-hash")]
 impl From<Error> for password_hash::Error {
     fn from(err: Error) -> Self {
         match err {
